Move a helper function into a new helpers/index.js file
This should set this up better for
when there are more helper functions.
