+++ tests/e2e/specs/ar-block.js
@@ -12,27 +12,12 @@ import uuid from 'uuid/v4';
 import {
 	clickButton,
 	createNewPost,
-	openGlobalBlockInserter,
-	pressKeyWithModifier,
 } from '@wordpress/e2e-test-utils';
 
 /**
- * Inserts a block from the block inserter, mainly copied from Gutenberg.
- *
- * @see https://github.com/WordPress/gutenberg/blob/56f912adc681ebd3a6fb6a17eb4cfcb2c0050f5b/packages/e2e-test-utils/src/insert-block.js
- *
- * @param {string} blockName The block name to search for.
+ * Internal dependencies
  */
-const insertBlockFromInserter = async ( blockName ) => {
-	await openGlobalBlockInserter();
-	await page.focus( '[placeholder="Search for a block"]' );
-	await pressKeyWithModifier( 'primary', 'a' );
-	await page.keyboard.type( blockName );
-	const insertButton = (
-		await page.$x( `//button//span[contains(text(), '${ blockName }')]` )
-	)[ 0 ];
-	await insertButton.click();
-};
+import { insertBlockFromInserter } from '../helpers';
 
 test( 'ar-viewer block', async () => {
 	const blockName = 'AR Viewer';
